third_party/blink/web_tests/fast/loader/loadInProgress-expected.txt

ALERT: Test passed: [object HTMLFormElement]
